updated skeleton list
