--- client/styles/layout/SkeletonList.module.css
@@ -1,3 +1,0 @@
-.skeletonList__skeleton {
-  width: 100%;
-}
